+++ src/lib/Documentary.js
@@ -1,0 +1,129 @@
+import { Replaceable, makeMarkers, makeCutRule, makePasteRule } from 'restream'
+import {
+  createTocRule, commentRule as stripComments, codeRe, innerCodeRe, linkTitleRe, linkRe,
+} from './rules'
+import tableRule, { tableRe } from './rules/table'
+import methodTitleRule, { methodTitleRe } from './rules/method-title'
+import treeRule from './rules/tree'
+import exampleRule from './rules/example'
+import forkRule from './rules/fork'
+import { getLink } from '.'
+import gifRule from './rules/gif'
+import typeRule from './rules/type'
+import badgeRule from './rules/badge'
+import typedefMdRule from './rules/typedef-md'
+import macroRule from './rules/macro'
+
+export default class Documentary extends Replaceable {
+  constructor({ toc } = {}) {
+    const tocRule = createTocRule(toc)
+
+    const {
+      table, methodTitle, code, innerCode, linkTitle,
+    } = makeMarkers({
+      table: tableRe,
+      methodTitle: methodTitleRe,
+      code: codeRe,
+      innerCode: innerCodeRe,
+      linkTitle: linkTitleRe,
+    })
+
+    /* below have ``` in them, therefore we want more control over handling them
+    * so that Replaceable does not confuse them with the code blocks.
+    */
+    const [cutCode, cutTable, cutMethodTitle, cutInnerCode] =
+      [code, table, methodTitle, innerCode, linkTitle].map((marker) => {
+        const rule = makeCutRule(marker)
+        return rule
+      })
+    const [insertCode, insertTable, insertMethodTitle, insertInnerCode] =
+      [code, table, methodTitle, innerCode, linkTitle].map((marker) => {
+        const rule = makePasteRule(marker)
+        return rule
+      })
+
+    super([
+      cutInnerCode, // don't want other rules being detected inside of inner code, including toc-titles
+      cutTable,
+      cutMethodTitle,
+      cutCode,
+      stripComments,
+
+      badgeRule,
+      treeRule,
+      ...[
+        exampleRule,
+        cutCode, // cut code again after inserting example
+      ],
+      forkRule,
+      tocRule,
+      gifRule,
+      typeRule,
+
+      insertTable,
+      typedefMdRule, // places a table hence just before table
+      macroRule, // macro is for the table
+      tableRule,
+
+      { // a hackish way to update types property tables to include links to seen types.
+        re: /\| _(\w+)_ \|/g,
+        replacement(match, name) {
+          if (!(name in this.types)) return match
+          return `| _[${name}](#${getLink(name)})_ |`
+        },
+      },
+      {
+        re: linkTitleRe,
+        replacement(match, title) {
+          const t = this.replaceInnerCode(title)
+          const link = getLink(t)
+          return `<a name="${link}">${t}</a>`
+        },
+      },
+      {
+        re: linkRe, // TODO implement links system
+        replacement(match, title) {
+          // TODO replace inner code as well
+          const link = getLink(title)
+          return `<a name="${link}">${title}</a>`
+        },
+      },
+      insertMethodTitle,
+      methodTitleRule,
+
+      insertCode,
+      insertInnerCode,
+      // those found inside of code blocks
+      insertTable,
+      insertMethodTitle,
+    ])
+
+    this._types = {}
+
+    this._innerCode = innerCode
+
+    this.on('types', types => {
+      types.forEach(this.addType.bind(this))
+    })
+  }
+  get innerCode() {
+    return this._innerCode
+  }
+  /**
+   * Replace a marked inner code with its actual value.
+   * @param {string} data
+   */
+  replaceInnerCode(data) {
+    const s = data.replace(this.innerCode.regExp, (m, i) => {
+      const val = this.innerCode.map[i]
+      return val
+    })
+    return s
+  }
+  addType(name) {
+    this.types[name] = true
+  }
+  get types() {
+    return this._types
+  }
+}