+++ src/lib/rules/example.js
@@ -1,6 +1,6 @@
 import { debuglog } from 'util'
 import { parse } from 'path'
-import { read } from '..'
+import { read, codeSurround } from '..'
 
 const LOG = debuglog('doc')
 
@@ -52,11 +52,12 @@ export const replacer = async (match, source, from, to, type) => {
       LOG('Example: %s', source)
     }
 
-    return `\`\`\`${getExt(type, source)}
-${ff.trim()}
-\`\`\``
-  } catch (err) {
-    LOG(err)
+    const lang = getExt(type, source)
+    const res = codeSurround(ff.trim(), lang)
+    return res
+  } catch ({ stack }) {
+    LOG('Could not read an example from %s.', source)
+    LOG(stack)
     return match
   }
 }