+++ test/make-rule-ts.js
@@ -1,7 +1,11 @@
 import { makeTestSuite } from 'zoroaster'
-import { Replaceable } from 'restream'
 import Context from './context'
 
+/**
+ * Run a replaceable stream against each item in the mask file.
+ * @param {string} path Path to the mask file
+ * @param {Rule|Rule[]} rule A rule or a set of rules to use.
+ */
 const makeRuleTestSuite = (path, rule) => {
   const ts = makeTestSuite(path, {
     /**
@@ -9,15 +13,18 @@ const makeRuleTestSuite = (path, rule) => {
      * @param {string} input
      * @param {Context} context
      */
-    async getResults(input, { catchment }) {
-      const rs = new Replaceable(rule)
-      rs.end(input)
-      const res = await catchment(rs)
+    async getResults(input, { replace }) {
+      const { res } = await replace(rule, input)
+      // console.log(res)
       return res
     },
     context: Context,
   })
   return ts
 }
+
+/**
+ * @typedef {import('restream').Rule} Rule
+ */
 
 export default makeRuleTestSuite