+++ test/spec/re/default.js
@@ -1,4 +1,4 @@
-import { ok, equal } from 'zoroaster/assert'
+import { equal, deepEqual } from 'zoroaster/assert'
 import SnapshotContext from 'snapshot-context'
 import Context from '../../context'
 import { tableRe } from '../../../src/lib/rules/table'
@@ -10,10 +10,9 @@ const T = {
     Context,
     SnapshotContext,
   ],
-  async 'matches the table with the re'({ table, cloneRe }) {
-    const re = cloneRe(tableRe)
-    const res = re.test(table)
-    ok(res)
+  async 'matches the table with the re'({ table, innerTable, mismatch }) {
+    const res = mismatch(tableRe, table, ['macro', 'content'])
+    deepEqual(res, [{ content: innerTable }])
   },
   async 'detects link titles'({ getMatches }) {
     const s = '[Hello World](t)'